Bug Fixed: Siren Not Stopping
Now, siren button works whether user switch to different activity or minimize the app or even close the app.

Used Singleton Class - 'MyMediaPlayer' to achieve the functionality.
