+++ app/build.gradle
@@ -38,4 +38,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     implementation "androidx.cardview:cardview:1.0.0"
     implementation 'com.google.android.gms:play-services-location:17.0.0'
+
+    //Gson
+    implementation 'com.google.code.gson:gson:2.8.6'
 }
