+++ app/src/main/java/com/example/she/siren.java
@@ -2,14 +2,15 @@ package com.example.she;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.SharedPreferences;
 import android.media.MediaPlayer;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-public class siren extends AppCompatActivity {
-    MediaPlayer player;
+import com.google.gson.Gson;
 
+public class siren extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -19,27 +20,10 @@ public class siren extends AppCompatActivity {
     }
 
     public void play(View view) {
-        if (player == null) {
-            player = MediaPlayer.create(this, R.raw.policesiren);
-            player.setLooping(true);
-            player.setOnCompletionListener(new MediaPlayer.OnCompletionListener() {
-                @Override
-                public void onCompletion(MediaPlayer mp) {
-                    stopPlayer();
-                }
-            });
-        }
-        player.start();
+        MyMediaPlayer.getInstance(this).play();
     }
 
     public void stop(View v) {
-        stopPlayer();
+        MyMediaPlayer.getInstance(this).stopPlayer();
     }
-    private void stopPlayer() {
-        if (player != null) {
-            player.release();
-            player = null;
-        }
-    }
-
 }
